Make In Memory Table Aggregator types more specific where possible (#3679)
Many aggregation types fell back to the general `Any` type where they could have used the type of input column - for example `First` of a column of integers is guaranteed to fit the `Integer` storage type, so it doesn't have to fall back to `Any`. This PR fixes that and adds a test that checks this.

+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/DateStorage.java
@@ -30,7 +30,7 @@ public class DateStorage extends SpecializedStorage<LocalDate> {
   }
 
   @Override
-  public long getType() {
+  public int getType() {
     return Type.DATE;
   }
 }
